% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_probs.R
\name{group.mem.probs}
\alias{group.mem.probs}
\title{Function for calculating group membership probabilities by chemical compositional distance using Mahalanobis distances and Hotellings T^2 statistic}
\usage{
group.mem.probs(x2.l, attr1.grp, grps)
}
\arguments{
\item{x2.l}{}

\item{attr1.grp}{}

\item{grps}{}
}
\description{
Function for calculating group membership probabilities by chemical compositional distance using Mahalanobis distances and Hotellings T^2 statistic
}
